+++ packages/@tinacms/cli/src/cmds/init/index.ts
@@ -7,17 +7,8 @@ import {
   ClientFile,
 } from './setup-files'
 import { TinaProvider, TinaProviderDynamic } from './setup-files/tinaProvider'
-import {
-  cmdText,
-  dangerText,
-  logText,
-  successText,
-  warnText,
-} from '../../utils/theme'
-import {
-  extendNextScripts,
-  generateGqlScript,
-} from '../../utils/script-helpers'
+import { logText, successText } from '../../utils/theme'
+import { extendNextScripts } from '../../utils/script-helpers'
 /**
 Copyright 2021 Forestry.io Holdings, Inc.
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -74,6 +65,11 @@ export const MIN_REACT_VERSION = '>=16.14.0'
 export async function checkDeps(ctx: any, next: () => void, options) {
   const bar = new Progress('Checking dependencies. :prog', 1)
 
+  if (!fs.existsSync(packageJSONPath)) {
+    throw new Error(
+      'No package.json Found. Please run tinacms init at the root of your app'
+    )
+  }
   const packageJSON = JSON.parse(
     (await fs.readFileSync(packageJSONPath)).toString()
   )
@@ -105,6 +101,9 @@ export const checkPackage = (packageJSON, packageName) => {
       }
     }
   )
+  if (!strippedVersion) {
+    throw new Error(`Please add ${packageName} to your project`)
+  }
   return checkVersion(strippedVersion)
 }
 
@@ -164,7 +163,6 @@ export async function installDeps(ctx: any, next: () => void, options) {
 }
 
 const baseDir = process.cwd()
-// TODO: should handle src folder here
 const packageJSONPath = p.join(baseDir, 'package.json')
 const blogContentPath = p.join(baseDir, 'content', 'posts')
 const blogPostPath = p.join(blogContentPath, 'HelloWorld.md')
@@ -246,18 +244,23 @@ export async function tinaSetup(_ctx: any, next: () => void, _options) {
   }
   logger.info('Adding a content folder... ✅')
   // 4. update the users package.json
-  const packagePath = p.join(baseDir, 'package.json')
-  const pack = JSON.parse(readFileSync(packagePath).toString())
-  const oldScripts = pack.scripts || {}
-  const newPack = JSON.stringify(
-    {
-      ...pack,
-      scripts: extendNextScripts(oldScripts),
-    },
-    null,
-    2
-  )
-  writeFileSync(packagePath, newPack)
+  if (!fs.existsSync(packageJSONPath)) {
+    throw new Error(
+      'No package.json Found. Please run tinacms init at the root of your app'
+    )
+  } else {
+    const pack = JSON.parse(readFileSync(packageJSONPath).toString())
+    const oldScripts = pack.scripts || {}
+    const newPack = JSON.stringify(
+      {
+        ...pack,
+        scripts: extendNextScripts(oldScripts),
+      },
+      null,
+      2
+    )
+    writeFileSync(packageJSONPath, newPack)
+  }
 
   // pages/admin.tsx
   const adminPath = p.join(pagesPath, 'admin.js')