+++ packages/@tinacms/cli/src/cmds/query-gen/genTypes.ts
@@ -11,12 +11,50 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { DocumentNode, GraphQLSchema, printSchema } from 'graphql'
+import { GraphQLSchema, printSchema } from 'graphql'
+import { TinaCloudSchema } from '@tinacms/schema-tools'
 
 import fs from 'fs-extra'
+import p from 'path'
 import { generateTypes } from '../../codegen'
-import { logText } from '../../utils/theme'
+import { logText, warnText } from '../../utils/theme'
 import { logger } from '../../logger'
+export const TINA_HOST = 'content.tinajs.io'
+const root = process.cwd()
+const generatedPath = p.join(root, '.tina', '__generated__')
+
+export async function genClient(
+  { tinaSchema }: { tinaSchema: TinaCloudSchema<false> },
+  next: () => void,
+  options
+) {
+  const branch = tinaSchema?.config?.branch
+  const clientId = tinaSchema?.config?.clientId
+  const token = tinaSchema?.config?.token
+
+  if (!branch || !clientId || !token) {
+    logger.warn(
+      // TODO: add link to docs
+      warnText('Client not configured properly. Skipping client generation.')
+    )
+    return next()
+  }
+
+  const apiURL = options.local
+    ? 'http://localhost:4001/graphql'
+    : `https://${TINA_HOST}/content/${clientId}/github/${branch}`
+
+  const clientPath = p.join(generatedPath, 'client.ts')
+  fs.writeFileSync(
+    clientPath,
+    `import { createClient } from "tinacms/dist/client";
+import { queries } from "./types";
+export const client = createClient({ url: '${apiURL}', token: '${token}', queries });
+export default client;
+  `
+  )
+  return next()
+}
 
 export async function genTypes(
   { schema }: { schema: GraphQLSchema },