+++ packages/@tinacms/graphql/CHANGELOG.md
@@ -1,5 +1,20 @@
 # tina-graphql
 
+## 0.61.3
+
+### Patch Changes
+
+- 0b5a8e6e7: Whitespace change to bump package
+
+## 0.61.2
+
+### Patch Changes
+
+- Updated dependencies [cf0f531a1]
+- Updated dependencies [b0dfc6205]
+  - @tinacms/datalayer@0.2.1
+  - @tinacms/schema-tools@0.0.8
+
 ## 0.61.1
 
 ### Patch Changes